backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
PyJWT
